refactor(components): split group state roles
